requirements.txt
kagglehub
pandas
# torch
# torchvision
scikit-learn
pillow
tqdm
logging
datetime
## pip install torch torchvision --index-url https://download.pytorch.org/whl/cu124